Adds wait, sleeper in registry and split tests
