+++ tests/public-api/requirements/_test.txt
@@ -4,32 +4,104 @@
 #
 #    pip-compile --output-file=requirements/_test.txt requirements/_test.in
 #
-attrs==20.3.0             # via pytest
-certifi==2020.12.5        # via httpx, osparc
-chardet==3.0.4            # via httpx
-contextvars==2.4          # via sniffio
-coverage==5.4             # via pytest-cov
-h11==0.9.0                # via httpcore
-httpcore==0.10.2          # via httpx
-httpx==0.14.3             # via -c requirements/../../../requirements/constraints.txt, -r requirements/_test.in
-idna==3.1                 # via rfc3986
-immutables==0.14          # via contextvars
-importlib-metadata==3.4.0  # via pluggy, pytest
-iniconfig==1.1.1          # via pytest
-git+https://github.com/ITISFoundation/osparc-simcore-python-client.git@catalog-solvers-api  # via -r requirements/_test.in
-packaging==20.9           # via pytest
-pluggy==0.13.1            # via pytest
-py==1.10.0                # via pytest
-pyparsing==2.4.7          # via packaging
-pytest-asyncio==0.14.0    # via -r requirements/_test.in
-pytest-cov==2.11.1        # via -r requirements/_test.in
-pytest==6.2.2             # via -r requirements/_test.in, pytest-asyncio, pytest-cov
-python-dateutil==2.8.1    # via osparc
-python-dotenv==0.15.0     # via -r requirements/_test.in
-rfc3986[idna2008]==1.4.0  # via httpx
-six==1.15.0               # via osparc, python-dateutil
-sniffio==1.2.0            # via httpcore, httpx
-toml==0.10.2              # via pytest
-typing-extensions==3.7.4.3  # via importlib-metadata
-urllib3==1.26.3           # via -c requirements/../../../requirements/constraints.txt, osparc
-zipp==3.4.0               # via importlib-metadata
+attrs==20.3.0
+    # via
+    #   jsonschema
+    #   pytest
+certifi==2020.12.5
+    # via
+    #   httpx
+    #   osparc
+    #   requests
+chardet==3.0.4
+    # via
+    #   httpx
+    #   requests
+contextvars==2.4
+    # via sniffio
+coverage==5.4
+    # via pytest-cov
+docker==4.4.1
+    # via -r requirements/_test.in
+h11==0.9.0
+    # via httpcore
+httpcore==0.10.2
+    # via httpx
+httpx==0.14.3
+    # via
+    #   -c requirements/../../../requirements/constraints.txt
+    #   -r requirements/_test.in
+idna==2.10
+    # via
+    #   requests
+    #   rfc3986
+immutables==0.14
+    # via contextvars
+importlib-metadata==3.4.0
+    # via
+    #   jsonschema
+    #   pluggy
+    #   pytest
+iniconfig==1.1.1
+    # via pytest
+jsonschema==3.2.0
+    # via -r requirements/_test.in
+git+https://github.com/ITISFoundation/osparc-simcore-python-client.git@catalog-solvers-api
+    # via -r requirements/_test.in
+packaging==20.9
+    # via pytest
+pluggy==0.13.1
+    # via pytest
+py==1.10.0
+    # via pytest
+pyparsing==2.4.7
+    # via packaging
+pyrsistent==0.17.3
+    # via jsonschema
+pytest-asyncio==0.14.0
+    # via -r requirements/_test.in
+pytest-cov==2.11.1
+    # via -r requirements/_test.in
+pytest==6.2.2
+    # via
+    #   -r requirements/_test.in
+    #   pytest-asyncio
+    #   pytest-cov
+python-dateutil==2.8.1
+    # via osparc
+python-dotenv==0.15.0
+    # via -r requirements/_test.in
+requests==2.25.1
+    # via docker
+rfc3986[idna2008]==1.4.0
+    # via httpx
+six==1.15.0
+    # via
+    #   docker
+    #   jsonschema
+    #   osparc
+    #   python-dateutil
+    #   tenacity
+    #   websocket-client
+sniffio==1.2.0
+    # via
+    #   httpcore
+    #   httpx
+tenacity==6.3.1
+    # via -r requirements/_test.in
+toml==0.10.2
+    # via pytest
+typing-extensions==3.7.4.3
+    # via importlib-metadata
+urllib3==1.26.3
+    # via
+    #   -c requirements/../../../requirements/constraints.txt
+    #   osparc
+    #   requests
+websocket-client==0.57.0
+    # via docker
+zipp==3.4.0
+    # via importlib-metadata
+
+# The following packages are considered to be unsafe in a requirements file:
+# setuptools
