+++ tests/public-api/requirements/_tools.txt
@@ -4,31 +4,78 @@
 #
 #    pip-compile --output-file=requirements/_tools.txt requirements/_tools.in
 #
-appdirs==1.4.4            # via black, virtualenv
-black==20.8b1             # via -r requirements/../../../requirements/devenv.txt
-bump2version==1.0.1       # via -r requirements/../../../requirements/devenv.txt
-cfgv==3.2.0               # via pre-commit
-click==7.1.2              # via black, pip-tools
-dataclasses==0.8          # via black
-distlib==0.3.1            # via virtualenv
-filelock==3.0.12          # via virtualenv
-identify==1.5.13          # via pre-commit
-importlib-metadata==3.4.0  # via -c requirements/_test.txt, pre-commit, virtualenv
-importlib-resources==5.1.0  # via pre-commit, virtualenv
-isort==5.7.0              # via -r requirements/../../../requirements/devenv.txt
-mypy-extensions==0.4.3    # via black
-nodeenv==1.5.0            # via pre-commit
-pathspec==0.8.1           # via black
-pip-tools==5.5.0          # via -r requirements/../../../requirements/devenv.txt
-pre-commit==2.10.0        # via -r requirements/../../../requirements/devenv.txt
-pyyaml==5.4.1             # via -c requirements/../../../requirements/constraints.txt, pre-commit
-regex==2020.11.13         # via black
-six==1.15.0               # via -c requirements/_test.txt, virtualenv
-toml==0.10.2              # via -c requirements/_test.txt, black, pre-commit
-typed-ast==1.4.2          # via black
-typing-extensions==3.7.4.3  # via -c requirements/_test.txt, black, importlib-metadata
-virtualenv==20.4.2        # via pre-commit
-zipp==3.4.0               # via -c requirements/_test.txt, importlib-metadata, importlib-resources
+appdirs==1.4.4
+    # via
+    #   black
+    #   virtualenv
+black==20.8b1
+    # via -r requirements/../../../requirements/devenv.txt
+bump2version==1.0.1
+    # via -r requirements/../../../requirements/devenv.txt
+cfgv==3.2.0
+    # via pre-commit
+click==7.1.2
+    # via
+    #   black
+    #   pip-tools
+dataclasses==0.8
+    # via black
+distlib==0.3.1
+    # via virtualenv
+filelock==3.0.12
+    # via virtualenv
+identify==1.5.13
+    # via pre-commit
+importlib-metadata==3.4.0
+    # via
+    #   -c requirements/_test.txt
+    #   pre-commit
+    #   virtualenv
+importlib-resources==5.1.0
+    # via
+    #   pre-commit
+    #   virtualenv
+isort==5.7.0
+    # via -r requirements/../../../requirements/devenv.txt
+mypy-extensions==0.4.3
+    # via black
+nodeenv==1.5.0
+    # via pre-commit
+pathspec==0.8.1
+    # via black
+pip-tools==5.5.0
+    # via -r requirements/../../../requirements/devenv.txt
+pre-commit==2.10.0
+    # via -r requirements/../../../requirements/devenv.txt
+pyyaml==5.4.1
+    # via
+    #   -c requirements/../../../requirements/constraints.txt
+    #   pre-commit
+regex==2020.11.13
+    # via black
+six==1.15.0
+    # via
+    #   -c requirements/_test.txt
+    #   virtualenv
+toml==0.10.2
+    # via
+    #   -c requirements/_test.txt
+    #   black
+    #   pre-commit
+typed-ast==1.4.2
+    # via black
+typing-extensions==3.7.4.3
+    # via
+    #   -c requirements/_test.txt
+    #   black
+    #   importlib-metadata
+virtualenv==20.4.2
+    # via pre-commit
+zipp==3.4.0
+    # via
+    #   -c requirements/_test.txt
+    #   importlib-metadata
+    #   importlib-resources
 
 # The following packages are considered to be unsafe in a requirements file:
 # pip
